App Task 입력 리팩토링
reducer를 분리함.
