+++ src/store.js
@@ -1,18 +1,5 @@
 import { createStore } from 'redux';
 
-const initialState = {
-  taskTitle: '',
-};
-
-export function reducer(state = initialState, action) {
-  if (action.type === 'Typing') {
-    return {
-      ...state,
-      title: action.payload,
-    };
-  }
-
-  return state;
-}
+import reducer from './reducer';
 
 export default createStore(reducer);
